+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-(empty)
+### Changed
+- Better info and error console logs, the error object is also printed out.
 
 ## [1.4.0] - 2020-07-05
 ### Changed